deduplicate as_list alias and fix header include order
