+++ include/camp/camp.hpp
@@ -4,11 +4,11 @@
 #include <array>
 #include <type_traits>
 
+#include "camp/size.hpp"
 #include "camp/defines.hpp"
 #include "camp/helpers.hpp"
 #include "camp/lambda.hpp"
-#include "camp/list/at.hpp"
-#include "camp/list/find_if.hpp"
+#include "camp/list.hpp"
 #include "camp/map.hpp"
 #include "camp/number.hpp"
 #include "camp/tuple.hpp"
@@ -19,8 +19,6 @@ namespace camp
 // Fwd
 template <typename... Ts>
 struct list;
-template <typename T>
-struct size;
 template <typename Seq>
 struct flatten;
 
@@ -205,33 +203,7 @@ namespace test
 }
 #endif
 
-namespace detail
-{
-  template <typename T>
-  struct _as_list;
-  template <template <typename...> class T, typename... Args>
-  struct _as_list<T<Args...>> {
-    using type = list<Args...>;
-  };
-  template <typename T, T... Args>
-  struct _as_list<int_seq<T, Args...>> {
-    using type = list<integral_constant<T, Args>...>;
-  };
-} /* detail */
-
-template <typename T>
-struct as_list_s : detail::_as_list<T>::type {
-};
-
-template <typename T>
-using as_list = typename as_list_s<T>::type;
-
 //// size
-template <typename... Args>
-struct size<list<Args...>> {
-  constexpr static idx_t value{sizeof...(Args)};
-  using type = num<sizeof...(Args)>;
-};
 
 #if defined(CAMP_TEST)
 namespace test
