+++ include/camp/size.hpp
@@ -1,0 +1,12 @@
+#ifndef __CAMP_SIZE_HPP
+#define __CAMP_SIZE_HPP
+
+namespace camp
+{
+
+template <typename T>
+struct size;
+
+} /* camp */
+
+#endif /* __CAMP_SIZE_HPP */
